requirements.txt
ultralytics==8.3.124
torch==2.0.1
torchvision==0.15.2
albumentations==1.3.1      
opencv-python==4.7.0.72
numpy==1.24.3
matplotlib==3.7.1
pyyaml==6.0
psutil==5.9.5
requests==2.31.0
Pillow==10.0.0
tqdm==4.66.1
pandas==1.5.3
seaborn==0.12.2
